Create user skill card component
